+++ RELEASE_NOTES.md
@@ -43,7 +43,8 @@ If you have problems upgrading to `7.0`, first try upgrading to `6.9`. It has a 
 * `EmojiCategoryTitle` is now available on all platforms.
 * `EmojiCategoryKeyboardMenu` is now available on all platforms.
 * `ExternalKeyboardContext` is now available on macOS and tvOS as well.
-* `KeyboardAction` gestures actions now use a `KeyboardActionTrigger` instead of a `KeyboardInputViewController`.  
+* `KeyboardAction` gesture actions are now available on all platforms.  
+* `KeyboardAction` gesture actions now use a `KeyboardActionTrigger` instead of a `KeyboardInputViewController`.  
 * `KeyboardActionTrigger` is a new protocol that lets us decouple actions from the input view controller.
 * `KeyboardInputViewController` implements `KeyboardActionTrigger` which gives it a bunch of new functions.
 * `GestureButton` is now available on watchOS 7.